+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 
 #### :bug: Bug fix
 
+- Make rewatch compile independent modules after an unrelated failure and recompile blocked dependents when a changed interface survives a failed implementation. https://github.com/rescript-lang/rescript/pull/8667
+
 #### :memo: Documentation
 
 #### :nail_care: Polish